% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mismatch_index.R
\name{compute_mismatch_index}
\alias{compute_mismatch_index}
\title{Compute mismatch index}
\usage{
compute_mismatch_index(data, as, by = NULL)
}
\arguments{
\item{as}{}
}
\value{

}
\description{
Compute mismatch index
}
